src/main/java/nomadrealms/game/GameStateSerializer.java
package nomadrealms.game;

import java.io.*;

public class GameStateSerializer {

    public static void serialize(GameState gameState, String filePath) throws IOException {
        try (ObjectOutputStream oos = new ObjectOutputStream(new FileOutputStream(filePath))) {
            oos.writeObject(gameState);
        }
    }

    public static GameState deserialize(String filePath) throws IOException, ClassNotFoundException {
        try (ObjectInputStream ois = new ObjectInputStream(new FileInputStream(filePath))) {
            return (GameState) ois.readObject();
        }
    }
}
